# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-Tenant SaaS POS System API — Phase 1 (Foundation)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [catalog]
      summary: List categories
      parameters:
        - name: includeInactive
          in: query
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [catalog]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [catalog]
      summary: Update a category
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteCategory
      tags: [catalog]
      summary: Deactivate a category
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deactivated
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /brands:
    get:
      operationId: listBrands
      tags: [catalog]
      summary: List brands
      parameters:
        - name: includeInactive
          in: query
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
    post:
      operationId: createBrand
      tags: [catalog]
      summary: Create a brand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /brands/{id}:
    patch:
      operationId: updateBrand
      tags: [catalog]
      summary: Update a brand
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteBrand
      tags: [catalog]
      summary: Deactivate a brand
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deactivated
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /colors:
    get:
      operationId: listColors
      tags: [catalog]
      summary: List colors
      parameters:
        - name: includeInactive
          in: query
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Colors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Color"
    post:
      operationId: createColor
      tags: [catalog]
      summary: Create a color
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ColorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Color"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /colors/{id}:
    patch:
      operationId: updateColor
      tags: [catalog]
      summary: Update a color
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ColorInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Color"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteColor
      tags: [catalog]
      summary: Deactivate a color
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deactivated
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /sizes:
    get:
      operationId: listSizes
      tags: [catalog]
      summary: List sizes
      parameters:
        - name: includeInactive
          in: query
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Sizes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Size"
    post:
      operationId: createSize
      tags: [catalog]
      summary: Create a size
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SizeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Size"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /sizes/{id}:
    patch:
      operationId: updateSize
      tags: [catalog]
      summary: Update a size
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SizeInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Size"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteSize
      tags: [catalog]
      summary: Deactivate a size
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deactivated
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        nameEn: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, isActive, createdAt]
    CategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        nameEn: { type: ["string", "null"] }
        isActive: { type: boolean }
    Brand:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        nameEn: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, isActive, createdAt]
    BrandInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        nameEn: { type: ["string", "null"] }
        isActive: { type: boolean }
    Color:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        nameEn: { type: ["string", "null"] }
        hex: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, isActive, createdAt]
    ColorInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        nameEn: { type: ["string", "null"] }
        hex: { type: ["string", "null"] }
        isActive: { type: boolean }
    Size:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        system: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, system, sortOrder, isActive, createdAt]
    SizeInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        system: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
...